Added Basic PoC C external function call and compiation script
Addition of a *.c file in the API folder will be compiled by GCC and
linked to the kernel. Additions to this system should work exactly as
before with the addition of an extern definition and a soon-to-come
global definition list for all api calls for use in new C api calls.